Implement monthly stats [closes #55]
This implements a fake model that aggregates statistics about badge
issuances and exposes that information at `/admin/stats`. Right now
the information given is low fidelity -- it only gives a total count of
badges per month, not a breakdown of total per badge. We could improve
upon this later if needed.