Frontend for selecting programs for badges.
Added a dropdown menu to the "new badge" & "edit badge" pages for
selecting a program. Added some text on the "show badge" page to display
which issuer and program the badge is associated with.

This commit also includes a lot of style fixes: whitespace fixes, using
`const` instead of `var` where appropriate, adding in missing
semicolons.