Frontend for issuing badges
This is a rather large commit containing the bare-bones necessities for
issuing badges manually. When a user signs in, we first check to see if
they are super user (as set by the environment variable). If not, we
check against each issuers' `accessList` to make a list of issuers that
a user has access to.

If a user has access to some issuers, we show the list of all the badges
related to those issuers (via their programs). The user is taken to a
screen where she can enter a list of email addresses and issue a badge
to each of them.