Implement multi-use claim codes
https://github.com/mozilla/openbadger/issues/83

In the original implementation once a claim code was used, it would be
unusable by anyone else unless it was manually released. This changeset
introduces the concept of multi-use claim codes which can be claimed
more than once.

Currently the number of times a multi-use claim code can
be used is unlimited, and we don't explicitly keep track of how many
times a code was used.