Implement tags for badges
Necessary for https://github.com/mozilla/openbadger/issues/86

We already had tags defined on the badges model, but we didn't have an
interface for actually setting the tags. This changeset provides that
interface and modifies the `create` and `update` handlers to expect tags.

This changeset also makes minor modifications to the "show badge" page:
the indicator that a badge should not be listed is now a sentence below
the badge name, and a section has been added for showing the existing
tags.