Add Badge method for getting recommendations
https://github.com/mozilla/openbadger/issues/86

We need to be able to make recommendations to users who are earning
badges about which badges to earn next. This takes the simplest
approach: given a badge, recommend other badges with those same tags to
users, filtering out the badges they have already earned. Currently the
recommendations are completely unsorted and unbiased.

Note, this changeset does *not* include the HTTP API, only the database
portion.