Add images to issuers.
This also does a lot of work to simplify things on the backend. Instead
of processing programs at the same time we try to process issuers, it is
a separate step. It's (arguably) a worse user experience on the
frontend, but A) the UX is in flux right now so we don't know what it's
going to look like anyway, and B) it's *way* simpler on the backend.