Added Badge#getRubricItems().
This method parses the criteria content by looking for bullet points
(lines that start with `*`) and converting them into rubric items
that match Aestimia's rubric item schema. If a bulleted line
contains the text `(optional)`, the corresponding rubric item is marked
as non-required; otherwise, it's required.

If no bulleted lines are found, the method returns an array with one
required rubric item, consisting of the text
`Satisfies the following criteria:\n` followed by the raw criteria text.

We still need to make API endpoints return the rubric information, but
hopefully this helps us get closer to a solution for #120.