Starting the recommendation system.
Right now we are doing the following:

* Filter earned badges
* Filter category-level badges
* Filter offline badges

We store this collection as `allBadges`, then

* Filter participation badges
* Filter badges that belong to categories where the user has already
  earned the category-level badge
* Filter badges that are "off-track", meaning, badges in categories
  where the user hasn't completed any other badges.

And we store this collection as `filtered`. If `filtered` is an empty
set, we shuffle `allBadges` and recommend that set.

There are a number of ways this can be improved which I have marked in a
TODO in models/badge.js