Initial implementation of ClaimCode.reservedFor.
This doesn't yet alter queries for available claim codes, e.g. the
getUnclaimedCodesTxt badge route. Nor does it implement email/webhook
callback when someone is issued a reserved claim code. It just
implements the notion of a reserved claim code in the badge/claim model,
along with a new Badge method, reserveAndEmail(), which creates a
reserved claim code, and changed the issueMany badge route to call it.