Fix console monkey patch for node v0.8
The way the `console` object is constructed changed between v0.8 and
v0.10. In the latest version of node the object is built from a
constructor and conveniently has _stdout and _stderr properties which
are references to streams which console will treat as stdout and
stderr. In v0.8 the console object is a singleton by virtue of the
methods being directly exported, and worse, the functions write to
process.stdout and process.stderr directly.